requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
itsdangerous==2.1.2
Werkzeug==2.3.7
